+++ CHANGELOG.md
@@ -11,10 +11,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 - Removed BC handler for deprecated `view` `_action`
-
-### Removed
 - The fallback mechanism that loads a default template when the template
 specified in a field description cannot be found was removed.
+- The Sonata Twig extension has been made final, you may no longer extend it to
+  build your own extension (which is deprecated by Twig anyway)
 
 ## [3.x]
 ### Added